refactoring: move collectUsedColumns from ExpressionAnalyzer to SyntaxAnalyzer
