+++ dbms/src/Interpreters/ExpressionAnalyzer.h
@@ -29,13 +29,8 @@ struct SyntaxAnalyzerResult;
 using SyntaxAnalyzerResultPtr = std::shared_ptr<const SyntaxAnalyzerResult>;
 
 /// ExpressionAnalyzer sources, intermediates and results. It splits data and logic, allows to test them separately.
-/// If you are not writing a test you probably don't need it. Use ExpressionAnalyzer itself.
 struct ExpressionAnalyzerData
 {
-    /// Original columns.
-    /// First, all available columns of the table are placed here. Then (when analyzing the query), unused columns are deleted.
-    NamesAndTypesList source_columns;
-
     /// If non-empty, ignore all expressions in  not from this list.
     NameSet required_result_columns;
 
@@ -58,15 +53,10 @@ struct ExpressionAnalyzerData
     /// Predicate optimizer overrides the sub queries
     bool rewrite_subqueries = false;
 
-    /// Columns will be added to block by join.
-    NamesAndTypesList columns_added_by_join;  /// Subset of analyzed_join.available_joined_columns
-
 protected:
-    ExpressionAnalyzerData(const NamesAndTypesList & source_columns_,
-                           const NameSet & required_result_columns_,
+    ExpressionAnalyzerData(const NameSet & required_result_columns_,
                            const SubqueriesForSets & subqueries_for_sets_)
-    :   source_columns(source_columns_),
-        required_result_columns(required_result_columns_),
+    :   required_result_columns(required_result_columns_),
         subqueries_for_sets(subqueries_for_sets_)
     {}
 };
@@ -102,7 +92,6 @@ public:
         const ASTPtr & query_,
         const SyntaxAnalyzerResultPtr & syntax_analyzer_result_,
         const Context & context_,
-        const NamesAndTypesList & additional_source_columns = {},
         const NameSet & required_result_columns_ = {},
         size_t subquery_depth_ = 0,
         bool do_global_ = false,
@@ -113,11 +102,6 @@ public:
 
     /// Get a list of aggregation keys and descriptions of aggregate functions if the query contains GROUP BY.
     void getAggregateInfo(Names & key_names, AggregateDescriptions & aggregates) const;
-
-    /** Get a set of columns that are enough to read from the table to evaluate the expression.
-      * Columns added from another table by JOIN are not counted.
-      */
-    Names getRequiredSourceColumns() const { return source_columns.getNames(); }
 
     /** These methods allow you to build a chain of transformations over a block, that receives values in the desired sections of the query.
       *
@@ -182,25 +166,21 @@ public:
     /// Create Set-s that we can from IN section to use the index on them.
     void makeSetsForIndex();
 
-    bool isRewriteSubqueriesPredicate() { return rewrite_subqueries; }
-
     bool hasGlobalSubqueries() { return has_global_subqueries; }
 
 private:
     ASTPtr query;
     const Context & context;
     const ExtractedSettings settings;
-    StoragePtr storage; /// The main table in FROM clause, if exists.
     size_t subquery_depth;
     bool do_global; /// Do I need to prepare for execution global subqueries when analyzing the query.
 
     SyntaxAnalyzerResultPtr syntax;
-    const AnalyzedJoin & analyzedJoin() const { return syntax->analyzed_join; }
 
-    /** Remove all unnecessary columns from the list of all available columns of the table (`columns`).
-      * At the same time, form a set of columns added by JOIN (`columns_added_by_join`).
-      */
-    void collectUsedColumns();
+    const StoragePtr & storage() const { return syntax->storage; } /// The main table in FROM clause, if exists.
+    const AnalyzedJoin & analyzedJoin() const { return syntax->analyzed_join; }
+    const NamesAndTypesList & sourceColumns() const { return syntax->required_source_columns; }
+    const NamesAndTypesList & columnsAddedByJoin() const { return syntax->columns_added_by_join; }
 
     /// Find global subqueries in the GLOBAL IN/JOIN sections. Fills in external_tables.
     void initGlobalSubqueriesAndExternalTables();
