+++ dbms/src/Storages/transformQueryForExternalDatabase.cpp
@@ -8,7 +8,6 @@
 #include <Parsers/ASTSelectQuery.h>
 #include <Parsers/ASTExpressionList.h>
 #include <Interpreters/SyntaxAnalyzer.h>
-#include <Interpreters/ExpressionAnalyzer.h>
 #include <Storages/transformQueryForExternalDatabase.h>
 #include <Storages/MergeTree/KeyCondition.h>
 
@@ -111,8 +110,7 @@ String transformQueryForExternalDatabase(
 {
     auto clone_query = query.clone();
     auto syntax_result = SyntaxAnalyzer(context).analyze(clone_query, available_columns);
-    ExpressionAnalyzer analyzer(clone_query, syntax_result, context);
-    const Names & used_columns = analyzer.getRequiredSourceColumns();
+    const Names used_columns = syntax_result->requiredSourceColumns();
 
     auto select = std::make_shared<ASTSelectQuery>();
